# src/immich_memories/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptView": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "id",
          "outcome",
          "phase",
          "reason",
          "run_id"
        ],
        "title": "AttemptView",
        "type": "object"
      },
...
      "Cleared": {
        "properties": {
          "name": {
            "enum": [
              "analysis",
              "video",
              "thumbnail",
              "preview"
            ],
            "title": "Name",
            "type": "string"
          },
          "removed": {
            "title": "Removed",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "removed"
        ],
        "title": "Cleared",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PhaseTiming": {
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "seconds": {
            "title": "Seconds",
            "type": "number"
          }
        },
        "required": [
          "name",
          "seconds",
          "errors"
        ],
        "title": "PhaseTiming",
        "type": "object"
      },
...
      "RunDetail": {
        "properties": {
          "child_output": {
            "title": "Child Output",
            "type": "boolean"
          },
          "clips_selected": {
            "title": "Clips Selected",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "date_range_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          },
          "date_range_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "delivery_status": {
            "title": "Delivery Status",
            "type": "string"
          },
          "film": {
            "title": "Film",
            "type": "boolean"
          },
          "has_cut": {
            "title": "Has Cut",
            "type": "boolean"
          },
          "memory_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Type"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/PhaseTiming"
            },
            "title": "Phases",
            "type": "array"
          },
          "preview_asset_ids": {
            "items": {
              "type": "string"
            },
            "title": "Preview Asset Ids",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "status",
          "source",
          "created_at",
          "memory_type",
          "date_range_start",
          "date_range_end",
          "preview_asset_ids",
          "film",
          "completed_at",
          "output_path",
          "delivery_status",
          "warnings",
          "phases",
          "clips_selected",
          "has_cut",
          "child_output"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunPage": {
        "properties": {
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "next_offset"
        ],
        "title": "RunPage",
        "type": "object"
      },
      "RunSuggestion": {
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "memory_key": {
            "title": "Memory Key",
            "type": "string"
          }
        },
        "required": [
          "memory_key"
        ],
        "title": "RunSuggestion",
        "type": "object"
      },
      "RunSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "date_range_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          },
          "date_range_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "film": {
            "title": "Film",
            "type": "boolean"
          },
          "memory_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Type"
          },
          "preview_asset_ids": {
            "items": {
              "type": "string"
            },
            "title": "Preview Asset Ids",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "source",
          "created_at",
          "memory_type",
          "date_range_start",
          "date_range_end",
          "preview_asset_ids",
          "film"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Immich Memories",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/automation/attempts/{attempt_id}": {
      "get": {
        "description": "Where an automation attempt stands, and the run it opened, even a failed one.",
        "operationId": "read_attempt_api_v1_automation_attempts__attempt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Attempt",
        "tags": [
          "suggestions"
        ]
      }
    },
...
    "/api/v1/caches/{name}/clear": {
      "post": {
        "description": "Empty one cache; the next run fills it again as it needs.",
        "operationId": "clear_cache_api_v1_caches__name__clear_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "enum": [
                "analysis",
                "video",
                "thumbnail",
                "preview"
              ],
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cleared"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Cache",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "description": "Runs newest first, each with the first pictures its saved cut plays.",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "completed",
                    "failed",
                    "running",
                    "cancelled",
                    "interrupted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "description": "One run as `runs show` reads it: outcome, output, delivery, warnings and phase timings.",
        "operationId": "read_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/child-output": {
      "get": {
        "description": "The retained output of an automatic run, credentials already removed when it was kept.",
        "operationId": "child_output_api_v1_runs__run_id__child_output_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Child Output",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/suggestions/run": {
      "post": {
        "description": "Start the candidate as `auto run` would: a lease, an attempt, the real CLI child.",
        "operationId": "run_suggestion_api_v1_suggestions_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSuggestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Suggestion",
        "tags": [
          "suggestions"
        ]
      }
    }
  }
}